chore: fixed imports after module restructure
