+++ frappe/tests/test_recorder.py
@@ -7,7 +7,7 @@ from __future__ import unicode_literals
 import unittest
 import frappe
 import frappe.recorder
-from .test_website import set_request
+from frappe.utils import set_request
 
 import sqlparse
