+++ frappe/tests/test_twofactor.py
@@ -5,7 +5,7 @@ from __future__ import unicode_literals
 import unittest, frappe, pyotp
 from frappe.auth import HTTPRequest
 from frappe.utils import cint
-from frappe.tests import set_request
+from frappe.utils import set_request
 from frappe.auth import validate_ip_address
 from frappe.twofactor import (should_run_2fa, authenticate_for_2factor, get_cached_user_pass,
 	two_factor_is_enabled_for_, confirm_otp_token, get_otpsecret_for_, get_verification_obj)
